+++ examples/webpack.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  entry: './app.js',
-  output: {
-    filename: 'bundle.js'
-  }
-};
-